Added a few more functions to the Camera class, it's probably finished until SDLloader is more fleshed out.
Added the ability to load generic .obj files into the mesh class. As of now, it ignores a good deal of the information contained in the .obj file, but it does load a good mesh that can be displayed and has lighting set up.

Added a (really poorly coded) example of how the camera functions can work and loading a mesh. It's not very clear, I'll fix it when I get the chance.