Good news everyone!
I've fixed the mesh class so that normals work. It will load normals from the file (which I've found to be quite bad, in general), or you can use mesh::calculateNormals to do it automatically. So load the file first, then calculate normals.

The problem before was a combination of three things:
1. A serious typo which stated that all of the normals and all of the vertices were in the same vector
2. Culling was not enabled (it is now)
3. Depth testing was not enabled (it is now)

Also, note that for now, all faces should be declared in the counterclockwise direction. I have edited GUIButton.cpp to reflect this.